requirements.txt
flask
pyserial
requests
datetime